requirements.txt
mypy==0.711
mypy-extensions==0.4.1
numpy==1.16.4
Pillow==6.0.0
typed-ast==1.4.0
